// src/utils/types/serviceOffering.ts
export interface IServiceOffering {
    name?: string;
    providedBy?: string;
    description?: string;
    createdAt?: Date;
    updatedAt?: Date;
    schema_version?: any;
    policy?: any;
    keywords?: any;
    category?: any;
    dataResources?: any;
    softwareResources?: any;
    location?: any;
    aggregationOf?: any;
    dependsOn?: any;
    termsAndConditions?: any;
}